backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.7.0
dj-database-url==2.3.0
python-dotenv==1.0.1
groq==0.18.0
google-genai
channels
PyMuPDF==1.25.3
easyocr==1.7.2
Pillow==11.1.0
opencv-python==4.11.0.86
numpy==2.2.3
psycopg[binary]==3.2.5
reportlab==4.2.5
feedparser
